Refactor iceqube to fit neatly in tasks. Replace asserts with TypeErrors.
