--- kolibri/core/tasks/iceqube/__init__.py
@@ -1,7 +1,0 @@
-from kolibri.core.tasks.iceqube.queue import Queue
-from kolibri.core.tasks.iceqube.worker import Worker
-
-
-__version__ = "0.4"
-
-__all__ = [__version__, Queue, Worker]
